feat: comments addressed
