+++ e2e/pages/pause-popup-page.ts
@@ -1,49 +1,60 @@
 import { Page, expect } from '@playwright/test';
 import { BasePage } from './base-page';
-import { Selectors } from '../constants/selectors';
 import { Timeouts } from '../constants/timeouts';
 
 export class PausePopupPage extends BasePage {
+  static override SELECTOR = '#pause-popup';
+
+  static SELECTORS = {
+    closeButton: '[data-click="close"]',
+    mapButton: '#pause-popup #map-button',
+    retryButton: '#pause-popup #retry-button',
+    confirmPopup: '#confirm-popup',
+    yesButton: '#yes-button',
+    cancelButton: '#cancel-button',
+  } as const;
+
   constructor(page: Page) {
     super(page);
   }
 
+  // container() inherited → getElement('#pause-popup')
+
   get popup() {
-    return this.page.locator(Selectors.pausePopup);
+    return this.getElement(PausePopupPage.SELECTOR);
   }
 
   get closeButton() {
-    return this.popup.locator(Selectors.pausePopupClose);
+    return this.popup.locator(PausePopupPage.SELECTORS.closeButton);
   }
 
   get mapButton() {
-    return this.page.locator(Selectors.pausePopupMapBtn);
+    return this.getElement(PausePopupPage.SELECTORS.mapButton);
   }
 
   get retryButton() {
-    return this.page.locator(Selectors.pausePopupRetryBtn);
+    return this.getElement(PausePopupPage.SELECTORS.retryButton);
   }
 
   get confirmPopup() {
-    return this.page.locator(Selectors.confirmPopup);
+    return this.getElement(PausePopupPage.SELECTORS.confirmPopup);
   }
 
   get yesButton() {
-    return this.page.locator(Selectors.yesButton);
+    return this.getElement(PausePopupPage.SELECTORS.yesButton);
   }
 
   get cancelButton() {
-    return this.page.locator(Selectors.cancelButton);
+    return this.getElement(PausePopupPage.SELECTORS.cancelButton);
   }
 
   async waitForPausePopup() {
-    // The popup has a 'show' CSS class when visible
     await this.page.waitForFunction(
       (sel) => {
         const el = document.querySelector(sel);
         return el && el.classList.contains('show');
       },
-      Selectors.pausePopup,
+      PausePopupPage.SELECTOR,
       { timeout: Timeouts.domUpdate },
     );
   }
@@ -77,7 +88,7 @@ export class PausePopupPage extends BasePage {
         const el = document.querySelector(sel);
         return el && el.classList.contains('show');
       },
-      Selectors.confirmPopup,
+      PausePopupPage.SELECTORS.confirmPopup,
       { timeout: Timeouts.domUpdate },
     );
   }
