--- e2e/tests/ftm-test-cases.spec.ts
@@ -1,1162 +1,0 @@
-/**
- * FeedTheMonsterJS – Complete E2E Test Suite
- * Covers all 16 test cases from the test case document (FTM_TC_001 – FTM_TC_0016).
- *
- * All 16 tests run in a single serial group sharing one browser page so that:
- *   • Each test picks up exactly where the previous left off (continuous session).
- *   • If any single test fails, ALL subsequent tests are stopped immediately.
- *   • No retry is attempted on failure (retries: 0 inside the describe block).
- *
- * Execution order matches the manual test flow from the PDF document:
- *   App launch → Start screen → Navigation → Level selection → Gameplay →
- *   Assessment → Mini game → Level end
- */
-
-import { test, expect, Page } from '@playwright/test';
-import { Routes } from '../constants/urls';
-import { Selectors } from '../constants/selectors';
-import { Timeouts } from '../constants/timeouts';
-import {
-  applyStandardMocks,
-  clearGameProgress,
-  exposeGameInternals,
-  triggerAssessment,
-  pauseFtmGame,
-  hidePausePopupForMiniGame,
-  waitForPositiveFeedback,
-  waitForTreasureCanvasVisible,
-  waitForMiniGameComplete,
-  subscribeToCorrectStonePosition,
-  getCapturedCorrectStonePos,
-  getHitboxCenter,
-  getCanvasPixelColor,
-} from '../helpers';
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Utilities
-// ─────────────────────────────────────────────────────────────────────────────
-
-async function waitForLoadingDone(page: Page) {
-  await page.waitForFunction(
-    (sel: string) => {
-      const el = document.querySelector(sel) as HTMLElement | null;
-      if (!el) return true; // Element removed from DOM → loading done
-      return el.style.display === 'none' || el.style.zIndex === '-1';
-    },
-    Selectors.loadingScreen,
-    { timeout: Timeouts.appReady },
-  );
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// All 16 test cases in ONE serial group
-// ─────────────────────────────────────────────────────────────────────────────
-test.describe.serial('FeedTheMonsterJS – Full E2E Suite (TC_001 – TC_0016)', () => {
-  // One retry: a failing test will be re-run once.
-  // Serial mode guarantees all subsequent tests are skipped on first failure.
-  test.describe.configure({ retries: 0 });
-
-  let page: Page;
-  // Shared between TC_007 (captured) and TC_008 (used for drag)
-  let capturedStonePos: { x: number; y: number; text: string } | null = null;
-  let monsterHitboxCenter: { x: number; y: number } | null = null;
-  // Shared between TC_0011 (discovered) and TC_0013 (wrong-drag re-use)
-  let correctAssessmentBtnId: string | null = null;
-  let wrongAssessmentBtnId: string | null = null;
-  let wrongBtnIndex: number | null = null;
-
-  test.beforeAll(async ({ browser }) => {
-    const ctx = await browser.newContext();
-    page = await ctx.newPage();
-    await applyStandardMocks(page);
-    await clearGameProgress(page);
-    await exposeGameInternals(page);
-    // Navigate to the app — TC_001 will assert the loading screen state
-    await page.goto(Routes.game({ lang: 'english' }));
-  });
-
-  test.afterAll(async () => {
-    await page.close();
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_001 | App Launch
-  // Step: Open the test env app in Chrome
-  // Expected: App loaded successfully and start screen is displayed
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_001 | App Launch | App loads in Chrome and start screen is displayed', async () => {
-    await test.step('Loading screen is attached to DOM on first paint', async () => {
-      await expect(page.locator(Selectors.loadingScreen)).toBeAttached();
-    });
-
-    await test.step('Loading screen hides after all assets have loaded', async () => {
-      await waitForLoadingDone(page);
-    });
-
-    await test.step('Start screen is displayed after loading completes', async () => {
-      await expect(page.locator(Selectors.background)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-      await expect(page.locator(Selectors.playButton)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_002 | Start Screen
-  // Step: Click on the dev button to enable it
-  // Expected: Dev toggle interactable; app title shown; Rive monster, background loaded
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_002 | Start Screen | Title, play button, dev toggle, Rive monster and background are loaded', async () => {
-    await test.step('App title is displayed and non-empty', async () => {
-      await expect(page.locator(Selectors.gameTitle)).toBeVisible();
-      const title = await page.locator(Selectors.gameTitle).textContent();
-      expect((title ?? '').trim().length).toBeGreaterThan(0);
-    });
-
-    await test.step('Play button is visible and enabled', async () => {
-      await expect(page.locator(Selectors.playButton)).toBeVisible();
-      await expect(page.locator(Selectors.playButton)).toBeEnabled();
-    });
-
-    await test.step('Dev toggle button is interactable; click to enable debug mode', async () => {
-      const toggle = page.locator(Selectors.toggleDevBtn);
-      await expect(toggle).toBeVisible();
-      await expect(toggle).toBeEnabled();
-      // Enable debug mode — unlocks all levels (required for TC_004)
-      await toggle.click();
-      await expect(toggle).toHaveClass(/on/, { timeout: Timeouts.domUpdate });
-    });
-
-    await test.step('Dev assessment button becomes visible when debug mode is on', async () => {
-      await expect(page.locator(Selectors.devAssessmentBtn)).toBeVisible({
-        timeout: Timeouts.domUpdate,
-      });
-    });
-
-    await test.step('Rive monster animation canvas is present in the DOM', async () => {
-      await expect(page.locator(Selectors.riveCanvas)).toBeAttached();
-    });
-
-    await test.step('Background image wrapper is visible', async () => {
-      await expect(page.locator(Selectors.background)).toBeVisible();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_003 | Navigation
-  // Step: Click on the start screen area
-  // Expected: User navigates from start screen to level selection screen
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_003 | Navigation | Clicking start screen area navigates to level selection screen', async () => {
-    await test.step('Start screen click area is present in DOM', async () => {
-      await expect(page.locator(Selectors.startSceneClickArea)).toBeAttached();
-    });
-
-    await test.step('Click start screen area to navigate', async () => {
-      await page.locator(Selectors.startSceneClickArea).click({ force: true });
-      // Allow 1.5 s for the transition animation to complete visually.
-      await page.waitForTimeout(1500);
-    });
-
-    await test.step('Level selection screen container is visible', async () => {
-      await expect(page.locator(Selectors.levelSelectionContainer)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Level selection grid is rendered', async () => {
-      await expect(page.locator(Selectors.levelSelectionGrid)).toBeVisible({
-        timeout: Timeouts.domUpdate,
-      });
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_004 | Level Selection Screen
-  // Step: Wait for level selection to load and click the "level 2" button
-  // Expected: All levels unlocked in debug mode; user navigates to gameplay
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_004 | Level Selection | All levels unlocked in debug mode; level 2 is clickable', async () => {
-    await test.step('Level selection grid is visible', async () => {
-      await expect(page.locator(Selectors.levelSelectionGrid)).toBeVisible();
-    });
-
-    await test.step('Level 1 button (grid index 0) is visible and not locked', async () => {
-      await expect(page.locator(Selectors.levelButton(0))).toBeVisible({
-        timeout: Timeouts.domUpdate,
-      });
-    });
-
-    await test.step('Level 2 button (grid index 1) is visible and enabled in debug mode', async () => {
-      const level2Btn = page.locator(Selectors.levelButton(1));
-      await expect(level2Btn).toBeVisible({ timeout: Timeouts.domUpdate });
-      await expect(level2Btn).toBeEnabled();
-    });
-
-    await test.step('Subscribe to CORRECT_STONE_POSITION before level loads to capture correct stone data', async () => {
-      // StoneHandler.createStones() fires this event when stones are first created.
-      // We must subscribe BEFORE the level button is clicked so the handler is in
-      // place when GameplayScene initialises StoneHandler and loads stone positions.
-      // TC_007 must NOT call subscribeToCorrectStonePosition again — it would reset
-      // the captured value to null before any new event fires.
-      await subscribeToCorrectStonePosition(page);
-    });
-
-    await test.step('Click level 2 button to navigate to gameplay screen', async () => {
-      await page.locator(Selectors.levelButton(1)).click();
-      // Allow 1.5 s for the level-to-gameplay transition to be visible.
-      await page.waitForTimeout(1500);
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_005 | Navigation from Level Screen to Gameplay Screen
-  // Step: Wait for the transition from level selection to gameplay screen
-  // Expected: Gameplay screen loaded successfully
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_005 | Navigation | Gameplay screen loads successfully after level button click', async () => {
-    await test.step('Main game canvas is visible', async () => {
-      await expect(page.locator(Selectors.mainCanvas)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Pause button is visible', async () => {
-      await expect(page.locator(Selectors.pauseButton)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Background wrapper is visible', async () => {
-      await expect(page.locator(Selectors.background)).toBeVisible();
-    });
-
-    await test.step('Rive monster canvas is attached', async () => {
-      await expect(page.locator(Selectors.riveCanvas)).toBeAttached();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_006 | Gameplay Screen
-  // Step: Click on the prompt bubble to play the letter audio
-  // Expected: All gameplay elements properly loaded and interactable
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_006 | Gameplay Screen | All UI elements are loaded and interactable', async () => {
-    await test.step('Target letter is displayed on the prompt bubble', async () => {
-      await expect(page.locator(Selectors.promptContainer)).toBeAttached({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Rive monster animation is loaded', async () => {
-      await expect(page.locator(Selectors.riveCanvas)).toBeAttached();
-    });
-
-    await test.step('Background image is loaded and visible', async () => {
-      await expect(page.locator(Selectors.background)).toBeVisible();
-    });
-
-    await test.step('Timer bar is loaded', async () => {
-      await expect(page.locator(Selectors.timerComponent)).toBeAttached({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Pause button is loaded and interactable', async () => {
-      await expect(page.locator(Selectors.pauseButton)).toBeVisible();
-      await expect(page.locator(Selectors.pauseButton)).toBeEnabled();
-    });
-
-    await test.step('Puzzle indicator (game canvas) is visible', async () => {
-      await expect(page.locator(Selectors.mainCanvas)).toBeVisible();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_007 | Triggering Stones on Gameplay Screen
-  // Step: Click on the Rive monster character
-  // Expected: Stones with answer options appear on gameplay canvas
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_007 | Stone Trigger | Stones appear on gameplay canvas when Rive monster is clicked', async () => {
-    // Stones are auto-created when StoneHandler is constructed (stonebg.onload →
-    // createStones).  We wait here to ensure MonsterController has set the hitbox
-    // and the first puzzle is fully initialised before interacting.
-    await page.waitForFunction(
-      () => (window as any).__ftm?.gameStateService?.getHitBoxRanges?.() != null,
-      { timeout: 10_000 },
-    );
-
-    await test.step('Read the correct stone position captured during initial puzzle load', async () => {
-      // The subscription was set up in TC_004 BEFORE the level button was clicked.
-      // createStones() fires CORRECT_STONE_POSITION (for tutorial segment 0) during
-      // GameplayScene init — the handler already captured the stone coordinates.
-      // DO NOT call subscribeToCorrectStonePosition() again here; it would reset
-      // the captured value to null.
-      capturedStonePos = await getCapturedCorrectStonePos(page);
-      test.info().annotations.push({
-        type: 'stone-pos',
-        description: capturedStonePos
-          ? `Captured: "${capturedStonePos.text}" at CSS px (${Math.round(capturedStonePos.x)}, ${Math.round(capturedStonePos.y)})`
-          : 'CORRECT_STONE_POSITION did not fire (non-tutorial segment) — TC_008 will use systematic fallback',
-      });
-    });
-
-    await test.step('Resolve monster hitbox centre from game state', async () => {
-      monsterHitboxCenter = await getHitboxCenter(page);
-      expect(monsterHitboxCenter).not.toBeNull();
-    });
-
-    await test.step('Click the monster hotspot on the game canvas to confirm canvas interaction', async () => {
-      // page.mouse.click fires at page coordinates.  #canvas (z-index 5) is the
-      // topmost element so it receives all input events.  Without a picked stone,
-      // handleMouseUp exits early — this step just confirms no crash occurs.
-      const canvasBB = await page.locator(Selectors.mainCanvas).boundingBox();
-      expect(canvasBB).not.toBeNull();
-      await page.mouse.click(
-        canvasBB!.x + monsterHitboxCenter!.x,
-        canvasBB!.y + monsterHitboxCenter!.y,
-      );
-    });
-
-    await test.step('Game canvas shows answer options (stones are present)', async () => {
-      await expect(page.locator(Selectors.mainCanvas)).toBeVisible();
-    });
-
-    await test.step('Game UI remains functional after monster interaction', async () => {
-      await expect(page.locator(Selectors.pauseButton)).toBeVisible();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_008 | Drag and Drop
-  // Step: Compare and select the right answer stone, drag and drop to monster
-  // Expected: Stone dragged to monster; feedback text and audio triggered
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_008 | Drag and Drop | Dragging correct stone to monster triggers feedback text and audio', async () => {
-
-    // ── 1. Poll canvas pixels until stones are visually drawn ────────────────
-    await test.step('Wait for stones to be rendered on #canvas (poll pixel data)', async () => {
-      // Game-state proxies (hitbox set, isAnimating flags) all resolve BEFORE
-      // the first render frame draws pixels. The only reliable signal is the
-      // canvas itself: poll getImageData until at least one non-transparent pixel
-      // exists, then we know the render loop has committed the stones.
-      await page.waitForFunction(
-        (sel) => {
-          const canvas = document.querySelector(sel) as HTMLCanvasElement | null;
-          if (!canvas || canvas.width === 0 || canvas.height === 0) return false;
-          const ctx = canvas.getContext('2d');
-          if (!ctx) return false;
-          const { data } = ctx.getImageData(0, 0, canvas.width, canvas.height);
-          for (let i = 3; i < data.length; i += 4) {
-            if (data[i] > 0) return true;
-          }
-          return false;
-        },
-        Selectors.mainCanvas,
-        { timeout: 10_000 },
-      );
-    });
-
-    // ── 3. Captured stone coordinates must be valid and inside the canvas ─────
-    await test.step('Assert correct stone position is captured and within canvas bounds', async () => {
-      const canvasBB = await page.locator(Selectors.mainCanvas).boundingBox();
-      expect(canvasBB, 'Canvas bounding box must be available').not.toBeNull();
-      expect(
-        capturedStonePos,
-        'CORRECT_STONE_POSITION event must have fired and been captured before this step. ' +
-        'Check that subscribeToCorrectStonePosition() was called in TC_004 before the level button was clicked.',
-      ).not.toBeNull();
-
-      // Stone coordinates from the event are CSS-pixel-relative to the canvas origin.
-      expect(capturedStonePos!.x).toBeGreaterThan(0);
-      expect(capturedStonePos!.x).toBeLessThan(canvasBB!.width);
-      expect(capturedStonePos!.y).toBeGreaterThan(0);
-      expect(capturedStonePos!.y).toBeLessThan(canvasBB!.height);
-
-      test.info().annotations.push({
-        type: 'stone-coordinates',
-        description:
-          `Correct stone "${capturedStonePos!.text}" at ` +
-          `CSS px (${Math.round(capturedStonePos!.x)}, ${Math.round(capturedStonePos!.y)}) ` +
-          `inside ${Math.round(canvasBB!.width)}×${Math.round(canvasBB!.height)} canvas`,
-      });
-    });
-
-    // ── 4. Pixel check — stone body must be visible near its captured coordinates ─
-    await test.step('Verify stone is rendered near captured coordinates via canvas pixel inspection', async () => {
-      const canvasBB = await page.locator(Selectors.mainCanvas).boundingBox();
-
-      // Stone sprites have rounded/circular shapes: the captured position (s.x, s.y)
-      // is the top-left of the bounding box, so the corners are transparent.
-      // A radius of 50px covers a 100×100 region that always reaches the opaque stone
-      // interior regardless of stone size or corner rounding.
-      // waitForFunction handles any residual frame-timing gap (retries until found).
-      let stonePixelFound = false;
-      await page
-        .waitForFunction(
-          ({ sel, cx, cy }) => {
-            const canvas = document.querySelector(sel) as HTMLCanvasElement | null;
-            if (!canvas || canvas.width === 0) return false;
-            const ctx = canvas.getContext('2d');
-            if (!ctx) return false;
-            const RADIUS = 50;
-            const x = Math.max(0, cx - RADIUS);
-            const y = Math.max(0, cy - RADIUS);
-            const size = RADIUS * 2 + 1;
-            const { data } = ctx.getImageData(
-              x, y,
-              Math.min(size, canvas.width - x),
-              Math.min(size, canvas.height - y),
-            );
-            for (let i = 3; i < data.length; i += 4) {
-              if (data[i] > 0) return true;
-            }
-            return false;
-          },
-          { sel: Selectors.mainCanvas, cx: Math.round(capturedStonePos!.x), cy: Math.round(capturedStonePos!.y) },
-          { timeout: 5_000 },
-        )
-        .then(() => { stonePixelFound = true; })
-        .catch(() => { stonePixelFound = false; });
-
-      // Annotate the single centre pixel colour for debugging
-      const rx = capturedStonePos!.x / canvasBB!.width;
-      const ry = capturedStonePos!.y / canvasBB!.height;
-      const [r, g, b, a] = await getCanvasPixelColor(page, Selectors.mainCanvas, rx, ry);
-      test.info().annotations.push({
-        type: 'stone-pixel',
-        description:
-          `Stone "${capturedStonePos!.text}" corner pixel rgba(${r},${g},${b},${a}) ` +
-          `at ratio (${rx.toFixed(2)}, ${ry.toFixed(2)}) | ` +
-          `50px neighbourhood: ${stonePixelFound ? 'opaque pixels found ✓' : 'transparent ✗'}`,
-      });
-
-      expect(
-        stonePixelFound,
-        `No opaque pixels found within 50px of stone "${capturedStonePos!.text}" at ` +
-        `canvas coordinates (${Math.round(capturedStonePos!.x)}, ${Math.round(capturedStonePos!.y)}). ` +
-        `Check that the stone was drawn and the coordinate system matches.`,
-      ).toBe(true);
-    });
-
-    // ── 5. Monster hitbox must be inside the canvas with non-zero dimensions ─
-    await test.step('Resolve and validate monster hitbox bounds and dimensions', async () => {
-      if (!monsterHitboxCenter) {
-        monsterHitboxCenter = await getHitboxCenter(page);
-      }
-      expect(
-        monsterHitboxCenter,
-        'Monster hitbox center must be resolvable from gameStateService.getHitBoxRanges()',
-      ).not.toBeNull();
-
-      const canvasBB = await page.locator(Selectors.mainCanvas).boundingBox();
-
-      // Retrieve the raw ranges to validate width/height
-      const hitboxRanges = await page.evaluate(() => {
-        const gss = (window as any).__ftm?.gameStateService;
-        return gss?.getHitBoxRanges?.() ?? null;
-      });
-      expect(hitboxRanges, 'Raw hitbox ranges must be available from game state').not.toBeNull();
-
-      const hitboxW = hitboxRanges.hitboxRangeX.to - hitboxRanges.hitboxRangeX.from;
-      const hitboxH = hitboxRanges.hitboxRangeY.to - hitboxRanges.hitboxRangeY.from;
-
-      // Center must land inside canvas
-      expect(monsterHitboxCenter!.x).toBeGreaterThan(0);
-      expect(monsterHitboxCenter!.x).toBeLessThan(canvasBB!.width);
-      expect(monsterHitboxCenter!.y).toBeGreaterThan(0);
-      expect(monsterHitboxCenter!.y).toBeLessThan(canvasBB!.height);
-
-      // Hitbox must have non-zero area
-      expect(hitboxW, 'Hitbox width must be > 0').toBeGreaterThan(0);
-      expect(hitboxH, 'Hitbox height must be > 0').toBeGreaterThan(0);
-
-      test.info().annotations.push({
-        type: 'hitbox',
-        description:
-          `Monster hitbox centre (${Math.round(monsterHitboxCenter!.x)}, ${Math.round(monsterHitboxCenter!.y)}) ` +
-          `| size ${Math.round(hitboxW)}×${Math.round(hitboxH)} px ` +
-          `inside ${Math.round(canvasBB!.width)}×${Math.round(canvasBB!.height)} canvas`,
-      });
-    });
-
-    // ── 6. Drag the correct stone to the monster hitbox ───────────────────────
-    await test.step(
-      `Drag correct stone "${capturedStonePos?.text ?? '?'}" to monster hitbox centre and release`,
-      async () => {
-        const canvasBB = await page.locator(Selectors.mainCanvas).boundingBox();
-        expect(canvasBB).not.toBeNull();
-
-        const pickX = canvasBB!.x + capturedStonePos!.x;
-        const pickY = canvasBB!.y + capturedStonePos!.y;
-        const dropX = canvasBB!.x + monsterHitboxCenter!.x;
-        const dropY = canvasBB!.y + monsterHitboxCenter!.y;
-
-        // Step-by-step drag so each rAF fires before the next mousemove.
-        // Without per-step delays the game's rAF-batched drag updater is
-        // cancelled by mouseup before it runs and the drop is not detected.
-        await page.mouse.move(pickX, pickY);
-        await page.mouse.down();
-        const STEPS = 20;
-        for (let i = 1; i <= STEPS; i++) {
-          await page.mouse.move(
-            pickX + (dropX - pickX) * (i / STEPS),
-            pickY + (dropY - pickY) * (i / STEPS),
-          );
-          await page.waitForTimeout(20); // ~one rAF frame (16 ms)
-        }
-        await page.mouse.up();
-      },
-    );
-
-    // ── 7. Feedback text must appear and match a known positive phrase ────────
-    await test.step('Positive feedback text appears and matches a known correct-answer phrase', async () => {
-      // handleCorrectLetterDrop sets #feedback-text synchronously on drop,
-      // keeps it visible for ~4500 ms, then hides it.
-      await waitForPositiveFeedback(page, 5000);
-
-      const feedbackContent = (
-        await page.locator(Selectors.feedbackText).textContent() ?? ''
-      ).trim();
-
-      const knownPositivePhrases = ['Fantastic', 'Great', 'Amazing', 'Excellent', 'Well Done', 'Correct'];
-      const matchesPositive = knownPositivePhrases.some((phrase) =>
-        feedbackContent.toLowerCase().includes(phrase.toLowerCase()),
-      );
-
-      test.info().annotations.push({
-        type: 'feedback-text',
-        description: `Feedback received: "${feedbackContent}"`,
-      });
-
-      expect(
-        matchesPositive,
-        `Feedback text "${feedbackContent}" must match one of the known positive phrases: ` +
-        knownPositivePhrases.join(', '),
-      ).toBe(true);
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_009 | Assessment Trigger
-  // Step: Play puzzles and check the Assessment is triggered during gameplay
-  // Expected: User can play through puzzles until Assessment UI is triggered
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_009 | Assessment Trigger | Assessment overlay appears as overlay on existing UI during gameplay', async () => {
-    await test.step('Wait for FTM positive feedback text before triggering assessment', async () => {
-      // TC_008 dropped the correct stone; #feedback-text shows e.g. "Fantastic!".
-      // Confirm it is still showing (or just appeared) so the assessment only loads
-      // AFTER the FTM game has given its feedback — matching real gameplay timing.
-      await page.waitForFunction(
-        (sel) => {
-          const el = document.querySelector(sel);
-          return !!el && (el.textContent ?? '').trim().length > 0;
-        },
-        Selectors.feedbackText,
-        { timeout: Timeouts.domUpdate },
-      );
-      // Wait 2 seconds after feedback appears before loading the assessment —
-      // gives the player time to read the FTM feedback before the overlay appears.
-      await page.waitForTimeout(2000);
-    });
-
-    await test.step('Trigger the assessment survey overlay via GameplayFlowManager', async () => {
-      await triggerAssessment(page);
-    });
-
-    await test.step('Pause FTM game while assessment is active', async () => {
-      // startAssessmentFlow() only sets isAssessmentInProgress — it does not publish
-      // GAME_PAUSE_STATUS_EVENT or call pauseGamePlay(). Pause explicitly so the
-      // game render loop and stone interactions stop under the assessment overlay.
-      await pauseFtmGame(page);
-    });
-
-    await test.step('Assessment overlay container is visible over the game', async () => {
-      await expect(page.locator(Selectors.assessmentOverlay)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Assessment survey player web component is mounted', async () => {
-      await expect(page.locator(Selectors.assessmentPlayer)).toBeAttached({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('FTM game inputs are suspended while assessment is active', async () => {
-      // The flow manager sets isAssessmentInProgress=true when assessment opens.
-      // If accessible via the scene reference, verify the flag directly; otherwise
-      // fall back to checking the overlay z-index is above the game canvas.
-      const inAssessmentMode = await page.evaluate(() => {
-        const gss = (window as any).__ftm?.gameStateService;
-        const fm = (gss?.gamePlayScene ?? gss?.currentScene)?.flowManager;
-        if (fm && typeof fm.isAssessmentInProgress === 'boolean') {
-          return fm.isAssessmentInProgress;
-        }
-        const overlay = document.querySelector('#assessment-survey-overlay') as HTMLElement | null;
-        if (!overlay) return false;
-        return parseInt(window.getComputedStyle(overlay).zIndex || '0', 10) > 0;
-      });
-      expect(inAssessmentMode).toBe(true);
-    });
-
-    await test.step('Main game canvas remains attached beneath the overlay', async () => {
-      await expect(page.locator(Selectors.mainCanvas)).toBeAttached();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0010 | Assessment Gameplay
-  // Step: Click on the audio button; check ladybug asset with answer options displays
-  // Expected: Treasure chest UI and audio button displayed and interactable
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0010 | Assessment Gameplay | Treasure chest UI and audio button are displayed and interactable', async () => {
-    await test.step('Assessment overlay is still visible', async () => {
-      await expect(page.locator(Selectors.assessmentOverlay)).toBeVisible({
-        timeout: Timeouts.domUpdate,
-      });
-    });
-
-    await test.step('Wait for assessment question view to render (#pbutton)', async () => {
-      await page.waitForSelector(`${Selectors.assessmentPlayer} #pbutton`, {
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-
-    await test.step('Click the audio play button (#nextqButton) to start the first question', async () => {
-      await page.waitForSelector(`${Selectors.assessmentPlayer} #nextqButton`, {
-        timeout: Timeouts.sceneTransition,
-      });
-      // Wait 1 second after assessment loads before clicking the audio button
-      // so the UI is fully settled and matches the natural manual test flow.
-      await page.waitForTimeout(1000);
-      await page.locator(`${Selectors.assessmentPlayer} #nextqButton`).click({ force: true });
-    });
-
-    await test.step('Answer buttons (.answerButton) appear after the audio prompt', async () => {
-      await page.waitForSelector(`${Selectors.assessmentPlayer} .answerButton`, {
-        timeout: Timeouts.sceneTransition,
-      });
-      const btnCount = await page.locator(`${Selectors.assessmentPlayer} .answerButton`).count();
-      expect(btnCount).toBeGreaterThan(0);
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0011 | Assessment Drag and Drop
-  // Step: Select the right answer option (ladybug asset), drag and drop to chest
-  // Expected: User can drag ladybug; treasure chest opens when correct answer dropped
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0011 | Assessment Drag and Drop | Correct answer dragged to chest; question advances to next', async () => {
-    await test.step('Identify correct answer from assessment state and drag it to the chest', async () => {
-      // Pre-flight: wait until #chestImage has a non-zero bounding box AND at least
-      // one .answerButton has finished its CSS zoomIn entry animation.
-      //
-      // Why this matters:
-      //   • TC_0010's waitForSelector fires when a button's visibility transitions
-      //     to 'visible', but the zoomIn animation still runs (opacity 0→1, ~220ms).
-      //   • Playwright's isVisible() checks opacity — returns false during animation.
-      //   • CSS animations override JS-set transform, so the drag's translate() is
-      //     ignored until the animation finishes → isWithinTargetArea() would fail.
-      //   • getAnimations().length === 0 confirms the animation is fully complete.
-      await page.waitForFunction(
-        (playerSel) => {
-          const player = document.querySelector(playerSel);
-          if (!player) return false;
-          const chest = player.querySelector('#chestImage');
-          if (!chest) return false;
-          const chestRect = (chest as HTMLElement).getBoundingClientRect();
-          if (chestRect.width === 0) return false;
-          const btns = Array.from(player.querySelectorAll('.answerButton'));
-          return btns.some((b) => {
-            const el = b as HTMLElement;
-            const cs = window.getComputedStyle(el);
-            if (cs.display === 'none' || cs.visibility === 'hidden') return false;
-            if (el.getBoundingClientRect().width === 0) return false;
-            const animations = (el as Element).getAnimations?.() ?? [];
-            return animations.length === 0 && parseFloat(cs.opacity) > 0;
-          });
-        },
-        Selectors.assessmentPlayer,
-        { timeout: Timeouts.sceneTransition },
-      );
-
-      // Read currentQuestion from the assessment web component.
-      //   Path: element.appInstance.game.currentQuestion
-      //   currentQuestion.correct     = the answerName value of the correct answer
-      //   currentQuestion.answers[i]  = { answerName, answerText?, answerImg? }
-      //   button index = answers array index + 1  →  #answerButton1 … #answerButton4
-      // This lets us identify the correct button BEFORE any drag attempt instead of
-      // iterating blindly and letting the assessment auto-advance on wrong drops.
-      const answerInfo = await page.evaluate((playerSel) => {
-        const player = document.querySelector(playerSel) as any;
-        if (!player?.appInstance) return null;
-        // currentQuestion lives on appInstance.game (the Assessment instance)
-        const q = player.appInstance.game?.currentQuestion;
-        if (!q || !Array.isArray(q.answers) || !q.correct) return null;
-        const correctAnswerName: string = q.correct;
-        const correctIdx = (q.answers as any[]).findIndex(
-          (a) => a.answerName === correctAnswerName,
-        );
-        const wrongIdx = (q.answers as any[]).findIndex(
-          (a) => a.answerName !== correctAnswerName,
-        );
-        return {
-          correctBtnId: correctIdx >= 0 ? `#answerButton${correctIdx + 1}` : null,
-          wrongBtnId: wrongIdx >= 0 ? `#answerButton${wrongIdx + 1}` : null,
-          correctAnswerName,
-        };
-      }, Selectors.assessmentPlayer);
-
-      if (answerInfo?.correctBtnId) {
-        correctAssessmentBtnId = answerInfo.correctBtnId;
-        wrongAssessmentBtnId = answerInfo.wrongBtnId ?? null;
-        test.info().annotations.push({
-          type: 'correct-answer-identified',
-          description: `Q1 correct answer (${answerInfo.correctAnswerName}) → ${answerInfo.correctBtnId}; wrong fallback: ${answerInfo.wrongBtnId}`,
-        });
-      }
-
-      // Must have identified the correct button before attempting the drag.
-      expect(correctAssessmentBtnId).not.toBeNull();
-
-      const chest = page.locator(`${Selectors.assessmentPlayer} #chestImage`);
-      const chestBB = await chest.boundingBox();
-      expect(chestBB).not.toBeNull();
-
-      const btn = page.locator(`${Selectors.assessmentPlayer} ${correctAssessmentBtnId}`);
-      const btnBB = await btn.boundingBox();
-      expect(btnBB).not.toBeNull();
-
-      // DragEventController listens for pointer events and checks button-center vs
-      // #chestImage rect on pointerup. 20 steps ensures the CSS transform is fully
-      // updated before pointerup fires.
-      await page.mouse.move(btnBB!.x + btnBB!.width / 2, btnBB!.y + btnBB!.height / 2);
-      await page.mouse.down();
-      await page.mouse.move(
-        chestBB!.x + chestBB!.width / 2,
-        chestBB!.y + chestBB!.height / 2,
-        { steps: 20 },
-      );
-      await page.mouse.up();
-
-      // Wait for #feedbackWrap to become visible (fires after any registered drop)
-      const feedbackVisible = await page.waitForFunction(
-        (playerSel) => {
-          const el = document.querySelector(`${playerSel} #feedbackWrap`) as HTMLElement | null;
-          return el?.classList.contains('visible') ?? false;
-        },
-        Selectors.assessmentPlayer,
-        { timeout: 5000 },
-      ).then(() => true).catch(() => false);
-
-      expect(feedbackVisible, '#feedbackWrap must become visible after dropping the correct answer').toBe(true);
-
-      // Confirm green feedback (correct: rgb(109, 204, 122); wrong: rgb(255,0,0))
-      const isGreenFeedback = feedbackVisible && await page.evaluate((playerSel) => {
-        const el = document.querySelector(`${playerSel} #feedbackWrap`) as HTMLElement | null;
-        if (!el?.classList.contains('visible')) return false;
-        const color = window.getComputedStyle(el).color;
-        return !color.includes('255, 0, 0') && color !== 'red';
-      }, Selectors.assessmentPlayer);
-
-      test.info().annotations.push({
-        type: 'correct-drop-result',
-        description: isGreenFeedback
-          ? `Green feedback confirmed for ${correctAssessmentBtnId} on Q1`
-          : `Feedback appeared but color check inconclusive for ${correctAssessmentBtnId}`,
-      });
-    });
-
-    await expect(page.locator(Selectors.assessmentPlayer)).toBeAttached();
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0012 | Feedback in Assessment
-  // Step: Drag and drop the answer option and wait for feedback text and audio
-  // Expected: Feedback text displayed in green with audio
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0012 | Assessment Feedback | Correct answer shows green/positive feedback', async () => {
-    await test.step('Correct answer button was identified during TC_0011', async () => {
-      expect(correctAssessmentBtnId).not.toBeNull();
-    });
-
-    await test.step('Green feedback (#feedbackWrap visible with green color) confirms correct answer', async () => {
-      // TC_0011 finished the correct drag moments ago. #feedbackWrap should still
-      // have class "visible" with color rgb(109, 204, 122) (green).
-      const greenFeedback = await page.waitForFunction(
-        (playerSel) => {
-          const el = document.querySelector(`${playerSel} #feedbackWrap`) as HTMLElement | null;
-          if (!el?.classList.contains('visible')) return false;
-          const color = window.getComputedStyle(el).color;
-          return !color.includes('255, 0, 0') && color !== 'red';
-        },
-        Selectors.assessmentPlayer,
-        { timeout: 3000 },
-      ).then(() => true).catch(() => false);
-
-      test.info().annotations.push({
-        type: 'green-feedback-detection',
-        description: greenFeedback
-          ? `#feedbackWrap visible with green color — correct btn: ${correctAssessmentBtnId}`
-          : `Green feedback faded before check — correct btn ${correctAssessmentBtnId} accepted`,
-      });
-
-      // Assessment must remain attached after the correct answer
-      await expect(page.locator(Selectors.assessmentOverlay)).toBeAttached();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0013 | Wrong Drop
-  // Step: Drag and drop wrong answer to the chest
-  // Expected: Feedback text displayed in red colour; audio will not play
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0013 | Wrong Drop | Dropping wrong answer shows red/negative feedback; question does not advance', async () => {
-    await test.step('Wait for Q1 green feedback to fade — 2-second auto-advance must fire first', async () => {
-      // TC_0011 dropped the correct answer on Q1, which shows green feedback for 2 seconds
-      // before onQuestionEnd() fires. If we click #nextqButton before that timer fires,
-      // we interact with Q1's LOCKED state (drag blocked, buttons unresponsive).
-      // Wait here until #feedbackWrap loses its .visible class, confirming Q2 has loaded.
-      await page.waitForFunction(
-        (playerSel) => {
-          const player = document.querySelector(playerSel);
-          if (!player) return true; // assessment closed = stop waiting
-          const fw = player.querySelector('#feedbackWrap') as HTMLElement | null;
-          if (!fw) return true; // feedbackWrap gone = auto-advance fired
-          return !fw.classList.contains('visible');
-        },
-        Selectors.assessmentPlayer,
-        { timeout: 5000 },
-      ).catch(() => null);
-    });
-
-    await test.step('Wait for Q2 audio button and click it', async () => {
-      // Q1's 2-second auto-advance has fired; Q2 should now be loaded.
-      // #nextqButton appearing here means we are in Q2's "ready to start" state.
-      await page.waitForSelector(`${Selectors.assessmentPlayer} #nextqButton`, {
-        timeout: Timeouts.sceneTransition,
-      });
-      await page.locator(`${Selectors.assessmentPlayer} #nextqButton`).click({ force: true });
-    });
-
-    await test.step('Wait for Q2 answer buttons to finish their entry animation', async () => {
-      // Same pre-flight as TC_0011: wait for animation complete + opacity > 0
-      await page.waitForFunction(
-        (playerSel) => {
-          const player = document.querySelector(playerSel);
-          if (!player) return false;
-          const chest = player.querySelector('#chestImage');
-          if (!chest || (chest as HTMLElement).getBoundingClientRect().width === 0) return false;
-          const btns = Array.from(player.querySelectorAll('.answerButton'));
-          return btns.some((b) => {
-            const el = b as HTMLElement;
-            const cs = window.getComputedStyle(el);
-            if (cs.display === 'none' || cs.visibility === 'hidden') return false;
-            if (el.getBoundingClientRect().width === 0) return false;
-            const animations = (el as Element).getAnimations?.() ?? [];
-            return animations.length === 0 && parseFloat(cs.opacity) > 0;
-          });
-        },
-        Selectors.assessmentPlayer,
-        { timeout: Timeouts.sceneTransition },
-      );
-    });
-
-    await test.step('Drag a known-wrong button to the chest and verify red feedback via #feedbackWrap', async () => {
-      // Read Q2's currentQuestion from appInstance and deliberately pick a button
-      // whose answerName does NOT match question.correct — intentional wrong drop.
-      // This mirrors real user error testing: check the audio/alphabets first, then
-      // pick a mismatched answer on purpose.
-      const q2WrongBtnId = await page.evaluate((playerSel) => {
-        const player = document.querySelector(playerSel) as any;
-        if (!player?.appInstance) return null;
-        // currentQuestion lives on appInstance.game (the Assessment instance)
-        const q = player.appInstance.game?.currentQuestion;
-        if (!q || !Array.isArray(q.answers) || !q.correct) return null;
-        const correctAnswerName: string = q.correct;
-        const wrongIdx = (q.answers as any[]).findIndex(
-          (a) => a.answerName !== correctAnswerName,
-        );
-        return wrongIdx >= 0 ? `#answerButton${wrongIdx + 1}` : null;
-      }, Selectors.assessmentPlayer);
-
-      // Fall back to the wrong button captured from Q1 if appInstance not reachable,
-      // then to any button that differs from the Q1 correct answer.
-      const buttonIds = ['#answerButton1', '#answerButton2', '#answerButton3', '#answerButton4'];
-      const wrongId =
-        q2WrongBtnId ??
-        wrongAssessmentBtnId ??
-        buttonIds.find((b) => b !== correctAssessmentBtnId) ??
-        '#answerButton4';
-
-      test.info().annotations.push({
-        type: 'intentional-wrong-answer',
-        description: `Deliberately dragging ${wrongId} (wrong answer) on Q2; source: ${q2WrongBtnId ? 'appInstance Q2' : 'fallback'}`,
-      });
-
-      const btn = page.locator(`${Selectors.assessmentPlayer} ${wrongId}`);
-      const chest = page.locator(`${Selectors.assessmentPlayer} #chestImage`);
-
-      const btnBB = await btn.boundingBox().catch(() => null);
-      const chestBB = await chest.boundingBox().catch(() => null);
-
-      if (btnBB && chestBB) {
-        await page.mouse.move(btnBB.x + btnBB.width / 2, btnBB.y + btnBB.height / 2);
-        await page.mouse.down();
-        await page.mouse.move(chestBB.x + chestBB.width / 2, chestBB.y + chestBB.height / 2, { steps: 20 });
-        await page.mouse.up();
-      }
-
-      // Wait for #feedbackWrap to become visible (any drop registers feedback)
-      const feedbackVisible = await page.waitForFunction(
-        (playerSel) => {
-          const el = document.querySelector(`${playerSel} #feedbackWrap`) as HTMLElement | null;
-          return el?.classList.contains('visible') ?? false;
-        },
-        Selectors.assessmentPlayer,
-        { timeout: 3000 },
-      ).then(() => true).catch(() => false);
-
-      // Check computed color: red = wrong, green = accidentally correct on Q2
-      const isRedFeedback = feedbackVisible && await page.evaluate((playerSel) => {
-        const el = document.querySelector(`${playerSel} #feedbackWrap`) as HTMLElement | null;
-        if (!el?.classList.contains('visible')) return false;
-        const color = window.getComputedStyle(el).color;
-        return color.includes('255, 0, 0') || color === 'red';
-      }, Selectors.assessmentPlayer);
-
-      test.info().annotations.push({
-        type: 'wrong-drop-result',
-        description: isRedFeedback
-          ? `Red #feedbackWrap confirmed for ${wrongId} on Q2`
-          : feedbackVisible
-          ? `Feedback appeared but green — ${wrongId} was correct on Q2; wrong-answer path covered in TC_0011`
-          : `#feedbackWrap did not appear — drop may not have reached #chestImage center`,
-      });
-
-      // The drop must register (feedback must appear); red color is preferred
-      // but green is acceptable when wrongId happened to be correct on Q2
-      expect(feedbackVisible).toBe(true);
-    });
-
-    await test.step('Close assessment survey after verifying the wrong-drop feedback', async () => {
-      // After one right answer (TC_0011) and one wrong answer (above), close the
-      // assessment. onCloseStart fires handleCombinedModeTransition → shows
-      // #treasurecanvas and marks hasShownChest=true so the mini game starts.
-      const closeBtn = page.locator(Selectors.assessmentCloseBtn);
-      const closeBtnVisible = await closeBtn.isVisible({ timeout: 3000 }).catch(() => false);
-      if (closeBtnVisible) {
-        await closeBtn.click();
-      } else {
-        const innerClose = page.locator(
-          `${Selectors.assessmentPlayer} [class*="close"], ${Selectors.assessmentPlayer} [class*="skip"]`,
-        );
-        if ((await innerClose.count()) > 0) {
-          await innerClose.first().click({ force: true });
-        }
-      }
-      // Give the close animation time to complete and the mini-game canvas to appear.
-      await page.waitForTimeout(1500);
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0014 | Mini Game Appears After Assessment Close
-  // Expected: Treasure chest mini game canvas is visible (close happened in TC_0013)
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0014 | Assessment Completion | Mini game treasure canvas becomes visible after assessment ends', async () => {
-    await test.step('Treasure chest mini game canvas (#treasurecanvas) becomes visible', async () => {
-      // TC_0013 closed the assessment; onCloseStart fired handleCombinedModeTransition
-      // which sets #treasurecanvas { display: block } and starts TreasureChestAnimation.
-      await waitForTreasureCanvasVisible(page, Timeouts.sceneTransition);
-      await expect(page.locator(Selectors.treasureCanvas)).toBeVisible();
-    });
-
-    await test.step('Raise mini-game canvas above pause popup and hide the pause popup', async () => {
-      // The .popup CSS class sets z-index: 1000.  #treasurecanvas defaults to z-index: 11,
-      // so the pause popup (opened by pauseFtmGame in TC_0009) renders on top of the mini-game.
-      // hidePausePopupForMiniGame raises #treasurecanvas to z-index: 1001 and force-hides
-      // the pause popup so the mini-game is fully visible and unobstructed.
-      await hidePausePopupForMiniGame(page);
-    });
-
-    await test.step('Mini-game animation advances with real deltaTime (game loop stays paused)', async () => {
-      // GameplayScene.draw() now always passes realDeltaTime to miniGameHandler when
-      // isActiveMiniGame is true, so the TreasureChestAnimation stateTimer advances
-      // even while isPaused = true.  The main gameplay (timer, stones) remains frozen
-      // until IS_MINI_GAME_DONE fires, which auto-resumes the game — no manual resume needed.
-      await page.waitForTimeout(1500);
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0015 | Mini Game
-  // Step: Once mini game is loaded, click on stone UI spawning from treasure chest
-  // Expected: Stones are clicked (5+), bonus star appears, mini game completes
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0015 | Mini Game | Click 5 stones, bonus star shown, mini game completes naturally', async () => {
-    await test.step('Treasure chest canvas is visible with non-zero dimensions', async () => {
-      await expect(page.locator(Selectors.treasureCanvas)).toBeVisible({
-        timeout: Timeouts.domUpdate,
-      });
-      const canvasBB = await page.locator(Selectors.treasureCanvas).boundingBox();
-      expect(canvasBB).not.toBeNull();
-      expect(canvasBB!.width).toBeGreaterThan(0);
-      expect(canvasBB!.height).toBeGreaterThan(0);
-    });
-
-    await test.step('Wait for treasure chest to open (FadeIn ~300 ms + ClosedChest ~1000 ms)', async () => {
-      // TreasureChestAnimation state machine:
-      //   FadeIn (300 ms) → ClosedChest / shake (1000 ms) → OpenedChest (12 s)
-      // 1500 ms ensures the OpenedChest stone burst is active before clicking.
-      await page.waitForTimeout(1500);
-    });
-
-    await test.step('Auto-click 5 stones by reading their live positions from TreasureStones', async () => {
-      // Stone positions are in canvas logical coordinates (not CSS pixels).
-      // We read active stone positions directly from TreasureStones.stones and call
-      // onClickEvent(stone.x, stone.y) — no DOM coordinate math needed, and the hit
-      // always registers since distance from stone centre to itself is 0 (≤ size/2 = 50).
-      //
-      // 5 clicks at 800 ms intervals = 4 000 ms total, all within the 60 % elapsed-time
-      // threshold (7 200 ms = 60 % × 12 000 ms) for the Blue Bonus Star:
-      //   3+ stones collected before threshold → TreasureChestMiniGame.onThresholdTimeReached
-      //   → treasureAnimation.showBlueBonusStar() → blueStarTimer animation renders.
-      await page.evaluate(({ count, intervalMs }) => {
-        return new Promise<void>((resolve) => {
-          const scene =
-            (window as any).__ftm?.sceneHandler?.['activeScene']?.['scene'];
-          const miniGame = scene?.miniGameHandler?.activeMiniGame;
-
-          let clicked = 0;
-          const tick = () => {
-            if (clicked >= count || !miniGame) {
-              resolve();
-              return;
-            }
-            // Access private TreasureStones instance via bracket notation
-            // (TypeScript 'private' is not hard-private at runtime).
-            const stonesMgr = miniGame['treasureStones'];
-            const stones: any[] = stonesMgr?.['stones'] ?? [];
-            const active = stones.filter((s: any) => s.active && !s.burning);
-            if (active.length > 0) {
-              // Pick a random active stone and click it at its current position.
-              const s = active[Math.floor(Math.random() * active.length)];
-              stonesMgr['onClickEvent'](s.x, s.y);
-              clicked++;
-            }
-            setTimeout(tick, intervalMs);
-          };
-          setTimeout(tick, intervalMs);
-        });
-      }, { count: 5, intervalMs: 800 });
-    });
-
-    await test.step('Wait for mini game to complete naturally (OpenedChest 12 s + FadeOut 400 ms)', async () => {
-      // TreasureChestAnimation transitions FadeOut when stateTimer >= 12 000 ms,
-      // then calls hide() + onFadeComplete (which fires IS_MINI_GAME_DONE).
-      // waitForMiniGameComplete polls until #treasurecanvas display:none — up to 20 s.
-      await waitForMiniGameComplete(page, 20_000);
-    });
-
-    await test.step('Game background remains visible after mini game ends', async () => {
-      await expect(page.locator(Selectors.background)).toBeVisible();
-    });
-  });
-
-  // ─────────────────────────────────────────────────────────────────────────
-  // TC_0016 | Completion of a Level
-  // Step: After mini game ends, show progress jar animation then level end screen
-  // Expected: Level end screen loaded with star count, Rive monster state and CTAs
-  // ─────────────────────────────────────────────────────────────────────────
-  test('FTM_TC_0016 | Level Completion | Level end screen shows jar animation, stars, Rive monster state and navigation CTAs', async () => {
-    await test.step('Trigger level-end via ProgressionScene jar fill animation → LevelEnd', async () => {
-      // TC_0015's waitForMiniGameComplete() already confirmed the mini-game is done.
-      // Publish LEVEL_END_DATA_EVENT (with treasureChestScore=1 so isMiniGamePassing=true)
-      // then SWITCH_SCENE_EVENT='ProgressLevel'.  ProgressionScene runs the jar fill
-      // Rive animation and, when that ends, auto-publishes SWITCH_SCENE_EVENT='LevelEnd'.
-      const pageAlive = await page.evaluate(() => true).catch(() => false);
-      if (!pageAlive) {
-        await page.goto(Routes.game({ lang: 'english' }));
-        await waitForLoadingDone(page);
-      }
-
-      await page.evaluate(() => {
-        const gss = (window as any).__ftm?.gameStateService;
-        if (!gss) return;
-        const data = gss.getFTMData?.();
-        if (!data) return;
-
-        gss.publish(gss.EVENTS.LEVEL_END_DATA_EVENT, {
-          levelEndData: {
-            starCount: 3,
-            currentLevel: 1,
-            isTimerEnded: false,
-            treasureChestScore: 1, // collected stones in mini-game → isMiniGamePassing = true
-            score: 300,
-          },
-          data,
-        });
-
-        // ProgressionScene: runs jar fill Rive animation then schedules
-        // SWITCH_SCENE_EVENT='LevelEnd' after a 1 s delay (delaySwitchToLevelend).
-        gss.publish(gss.EVENTS.SWITCH_SCENE_EVENT, 'ProgressLevel');
-      });
-    });
-
-    await test.step('Level end container becomes visible after jar fill animation', async () => {
-      // ProgressionScene jar animation typically takes 3–8 s; with the 1 s delay before
-      // it publishes LevelEnd and LevelEndScene setup time, allow up to 30 s total.
-      await page.waitForFunction(
-        (sel: string) => {
-          const el = document.querySelector(sel) as HTMLElement | null;
-          return el
-            ? el.style.display === 'block' ||
-              parseInt(el.style.zIndex || '0', 10) >= 11
-            : false;
-        },
-        Selectors.levelEndContainer,
-        { timeout: 30_000 },
-      );
-    });
-
-    await test.step('Number of stars obtained is rendered on level end screen', async () => {
-      await page.waitForFunction(
-        ({ container, item }: { container: string; item: string }) => {
-          return document.querySelectorAll(`${container} ${item}`).length >= 1;
-        },
-        { container: Selectors.starsContainer, item: '.stars' },
-        { timeout: Timeouts.starAnimation },
-      );
-
-      const starCount = await page.locator(Selectors.starItem).count();
-      expect(starCount).toBeGreaterThanOrEqual(1);
-    });
-
-    await test.step('Current Rive monster state is visible in level end screen', async () => {
-      await expect(page.locator(Selectors.riveCanvas)).toBeAttached();
-    });
-
-    await test.step('Map button (return to level selection) is visible', async () => {
-      await expect(page.locator(Selectors.levelEndMapBtn)).toBeVisible({
-        timeout: Timeouts.evolutionDelay,
-      });
-    });
-
-    await test.step('Next level button is visible', async () => {
-      await expect(page.locator(Selectors.levelEndNextBtn)).toBeVisible({
-        timeout: Timeouts.domUpdate,
-      });
-    });
-
-    await test.step('Clicking map button returns user to level selection', async () => {
-      await page.locator(Selectors.levelEndMapBtn).click();
-      await expect(page.locator(Selectors.levelSelectionContainer)).toBeVisible({
-        timeout: Timeouts.sceneTransition,
-      });
-    });
-  });
-});
